Added database password handling
Have added a simple dialog which allows the database password to be
entered. Once entered it is validated by connecting to the DB and saved
locally as a .scrote file (or at least on my pc -it might not make the
'config' folder structure yet). I have changed gitIgnore so that it wont
sync .scrote files which will allow us to keep local copies of the
password without muggins mcHacker getting at the DB.